[refactor] Extract common code into MultiMountSourceMixin
Extract shared functionality between LinkResolutionUnionMountSource and
UnionMountSource into a new MultiMountSourceMixin class. This includes:
- open(), read() methods for delegating file operations
- list_xattr(), get_xattr() for extended attribute handling
- get_mount_source() for mount source delegation
- is_immutable(), statfs() for filesystem metadata
- __enter__(), __exit__() for context management
